New script scaffolding; improved unit tests
