+++ scripts/whitelist-fiscal-code-script/src/app/__test__/config.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from 'vitest';
 
-import { checkConfig } from './config';
+import { checkConfig } from '../config';
 
 vi.mock('winston', () => ({
   createLogger: vi.fn().mockImplementation(() => ({
@@ -48,16 +48,9 @@ describe('checkConfig()', () => {
     process.env.SLEEP_TIME_BETWEEN_REQUESTS_MS = '1000';
     process.env.REQUEST_TIMEOUT_MS = '10000';
 
-    try {
-      checkConfig();
-    } catch (error) {
-      expect((error as Error).message).toContain(
-        'DATABASE_CONNECTION_STRING is required',
-      );
-      return;
-    }
-
-    expect(true).toStrictEqual(false);
+    expect(() => checkConfig()).toThrow(
+      'DATABASE_CONNECTION_STRING is required',
+    );
   });
 
   it('should fail validation when DATABASE_CONNECTION_STRING is an empty string', () => {
@@ -67,16 +60,9 @@ describe('checkConfig()', () => {
     process.env.SLEEP_TIME_BETWEEN_REQUESTS_MS = '1000';
     process.env.REQUEST_TIMEOUT_MS = '10000';
 
-    try {
-      checkConfig();
-    } catch (error) {
-      expect((error as Error).message).toContain(
-        'DATABASE_CONNECTION_STRING is not allowed to be empty',
-      );
-      return;
-    }
-
-    expect(true).toStrictEqual(false);
+    expect(() => checkConfig()).toThrow(
+      'DATABASE_CONNECTION_STRING is not allowed to be empty',
+    );
   });
 
   it('should fail validation when SLEEP_TIME_BETWEEN_REQUESTS_MS is of another type', () => {
@@ -86,16 +72,9 @@ describe('checkConfig()', () => {
     process.env.SLEEP_TIME_BETWEEN_REQUESTS_MS = 'wrong-type'; // string, instead of number
     process.env.REQUEST_TIMEOUT_MS = '10000';
 
-    try {
-      checkConfig();
-    } catch (error) {
-      expect((error as Error).message).toContain(
-        'SLEEP_TIME_BETWEEN_REQUESTS_MS must be a number',
-      );
-      return;
-    }
-
-    expect(true).toStrictEqual(false);
+    expect(() => checkConfig()).toThrow(
+      'SLEEP_TIME_BETWEEN_REQUESTS_MS must be a number',
+    );
   });
 
   it('should fail validation when REQUEST_TIMEOUT_MS is of another type', () => {
@@ -105,16 +84,7 @@ describe('checkConfig()', () => {
     process.env.SLEEP_TIME_BETWEEN_REQUESTS_MS = '1000';
     process.env.REQUEST_TIMEOUT_MS = 'wrong-type'; // string, instead of number
 
-    try {
-      checkConfig();
-    } catch (error) {
-      expect((error as Error).message).toContain(
-        'REQUEST_TIMEOUT_MS must be a number',
-      );
-      return;
-    }
-
-    expect(true).toStrictEqual(false);
+    expect(() => checkConfig()).toThrow('REQUEST_TIMEOUT_MS must be a number');
   });
 
   it('should fail validation when DATABASE_NAME is an empty string', () => {
@@ -124,16 +94,9 @@ describe('checkConfig()', () => {
     process.env.SLEEP_TIME_BETWEEN_REQUESTS_MS = '1000';
     process.env.REQUEST_TIMEOUT_MS = '10000';
 
-    try {
-      checkConfig();
-    } catch (error) {
-      expect((error as Error).message).toContain(
-        'DATABASE_NAME is not allowed to be empty',
-      );
-      return;
-    }
-
-    expect(true).toStrictEqual(false);
+    expect(() => checkConfig()).toThrow(
+      'DATABASE_NAME is not allowed to be empty',
+    );
   });
 
   it('should fail validation when DATABASE_CONTAINER_NAME is an empty string', () => {
@@ -143,15 +106,8 @@ describe('checkConfig()', () => {
     process.env.SLEEP_TIME_BETWEEN_REQUESTS_MS = '1000';
     process.env.REQUEST_TIMEOUT_MS = '10000';
 
-    try {
-      checkConfig();
-    } catch (error) {
-      expect((error as Error).message).toContain(
-        'DATABASE_CONTAINER_NAME is not allowed to be empty',
-      );
-      return;
-    }
-
-    expect(true).toStrictEqual(false);
+    expect(() => checkConfig()).toThrow(
+      'DATABASE_CONTAINER_NAME is not allowed to be empty',
+    );
   });
 });
