+++ scripts/whitelist-fiscal-code-script/src/get-argv-param.ts
