+++ scripts/whitelist-fiscal-code-script/src/get-logger.ts
